+++ CHANGELOG.md
@@ -17,7 +17,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `demo1()`
   - `demo2()`
 - envelopes module: `ADSR_envelope()` function
-- signals module: `add_karplus_strong()` for plucked strings
+- signals module: 
+  - `add_karplus_strong()` for plucked strings
+  - `add_square_wave`
 - sound_effects module: 
   - `apply_delay_effect()`
   - `apply_fuzz_effect()`